# docs/swagger.yaml
definitions:
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      errors:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
  description: fagin服务.
  termsOfService: https://github.com/Vfuryx/fagin
  title: server
  version: "2.0"
paths:
  /:
    get:
      produces:
      - text/html
      responses:
        "200":
          description: ""
      summary: 首页
      tags:
      - web
  /admin/api/v1/admins/:
    post:
      consumes:
      - application/json
      description: 管理员列表接口
      parameters:
      - description: 用户名称
        in: query
        maxLength: 64
        name: username
        type: string
      - description: 手机号码
        in: query
        maxLength: 11
        minLength: 11
        name: phone
        type: string
      - description: 状态
        enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '正确 {"code":0,"message":"OK","data":{"token":"XXXXX"}} <br/>
            错误 {"code": 20102,"message": "找不到用户"} <br/> "{code": 20105, "message":
            "密码不正确"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 管理员列表
      tags:
      - 总后台接口
  /api/v1/login:
    post:
      consumes:
      - multipart/form-data
      description: 用户登录
      parameters:
      - description: 用户邮箱地址
        in: formData
        maxLength: 16
        minLength: 6
        name: email
        required: true
        type: string
      - description: 用户密码
        in: formData
        maxLength: 32
        minLength: 6
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '正确 {"code":0,"message":"OK","data":{"token":"XXXXX"}} <br/>
            错误 {"code": 20102,"message": "找不到用户"} <br/> "{code": 20105, "message":
            "密码不正确"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 登录接口
      tags:
      - auth
  /api/video/play/{id}:
    get:
      consumes:
      - multipart/form-data
      description: 播放视频
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '成功返回视频溜 <br/> 失败返回 {"code": 10005,"message": "打开文件失败"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 播放视频接口
      tags:
      - video
swagger: "2.0"